--- benchmark/bench_test.go
@@ -1,66 +1,0 @@
-package benchmark
-
-import (
-	"fmt"
-	bitcask "github.com/XiXi-2024/xixi-bitcask-kv"
-	"github.com/XiXi-2024/xixi-bitcask-kv/utils"
-	"github.com/stretchr/testify/assert"
-	"math/rand"
-	"os"
-	"testing"
-	"time"
-)
-
-var db *bitcask.DB
-
-func init() {
-	// 初始化用于基准测试的 DB 实例
-	options := bitcask.DefaultOptions
-	dir, _ := os.MkdirTemp("", "bitcask-go-bench")
-	options.DirPath = dir
-	fmt.Println(dir)
-
-	var err error
-	db, err = bitcask.Open(options)
-	if err != nil {
-		panic(err)
-	}
-}
-
-func Benchmark_Put(b *testing.B) {
-	b.ResetTimer()
-	b.ReportAllocs()
-
-	for i := 0; i < b.N; i++ {
-		err := db.Put(utils.GetTestKey(i), utils.RandomValue(1024))
-		assert.Nil(b, err)
-	}
-}
-
-func Benchmark_Get(b *testing.B) {
-	for i := 0; i < 10000; i++ {
-		err := db.Put(utils.GetTestKey(i), utils.RandomValue(1024))
-		assert.Nil(b, err)
-	}
-
-	rand.Seed(time.Now().UnixNano())
-	b.ResetTimer()
-	b.ReportAllocs()
-	for i := 0; i < b.N; i++ {
-		_, err := db.Get(utils.GetTestKey(rand.Int()))
-		if err != nil && err != bitcask.ErrKeyNotFound {
-			b.Fatal(err)
-		}
-	}
-}
-
-func Benchmark_Delete(b *testing.B) {
-	b.ResetTimer()
-	b.ReportAllocs()
-
-	rand.Seed(time.Now().UnixNano())
-	for i := 0; i < b.N; i++ {
-		err := db.Delete(utils.GetTestKey(rand.Int()))
-		assert.Nil(b, err)
-	}
-}
